% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeSIR.R
\name{ridgeRes}
\alias{ridgeRes}
\alias{summary.ridgeRes}
\alias{print.ridgeRes}
\alias{ridgeRes-class}
\title{Print ridgeRes object}
\usage{
\method{summary}{ridgeRes}(object, ...)

\method{print}{ridgeRes}(x, ...)
}
\arguments{
\item{object}{a \code{ridgeRes} object}

\item{...}{not used}

\item{x}{a \code{ridgeRes} object}
}
\description{
Print a summary of the result of \code{\link{ridgeSIR}} (
\code{ridgeRes} object)
}
\seealso{
\code{\link{ridgeSIR}}
}
\author{
{Victor Picheny, \email{victor.picheny@inrae.fr}\cr 
Remi Servien, \email{remi.servien@inrae.fr}\cr
Nathalie Vialaneix, \email{nathalie.vialaneix@inrae.fr}}
}
