% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseSIR.R
\name{SISIRres}
\alias{SISIRres}
\alias{print.SISIRres}
\alias{summary.SISIRres}
\title{Print SISIRres object}
\usage{
\method{summary}{SISIRres}(object, ...)

\method{print}{SISIRres}(x, ...)
}
\arguments{
\item{object}{a \code{SISIRres} object}

\item{...}{not used}

\item{x}{a \code{SISIRres} object}
}
\description{
Print a summary of the result of \code{\link{SISIRres}} (
\code{SISIRres} object)
}
\author{
{Victor Picheny, \email{victor.picheny@inra.fr}

Remi Servien, \email{remi.servien@inra.fr}

Nathalie Villa-Vialaneix, \email{nathalie.villa-vialaneix@inra.fr}}
}
\seealso{
\code{\link{SISIR}}
}

