% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseSIR.R
\name{sparseRes}
\alias{print.sparseRes}
\alias{sparseRes}
\alias{summary.sparseRes}
\title{Print sparseRes object}
\usage{
\method{summary}{sparseRes}(object, ...)

\method{print}{sparseRes}(x, ...)
}
\arguments{
\item{object}{a \code{sparseRes} object}

\item{...}{not used}

\item{x}{a \code{sparseRes} object}
}
\description{
Print a summary of the result of \code{\link{sparseSIR}} (
\code{sparseRes} object)
}
\author{
{Victor Picheny, \email{victor.picheny@inra.fr}

Remi Servien, \email{remi.servien@inra.fr}

Nathalie Villa-Vialaneix, \email{nathalie.villa-vialaneix@inra.fr}}
}
\seealso{
\code{\link{sparseSIR}}
}

